+++ AutomaticRoleCheck.lua
@@ -2,6 +2,7 @@ AutomaticRoleCheck = {
     FirstCheck = true,
     Defaults = {
       Enabled = true,
+      DisableWhilstAFK = false,
       DisableOnce = false,
       DisableOnceOnLogin = false,
     }
@@ -12,11 +13,17 @@ if not AutomaticRoleCheck_Options then
 end
 
 AutomaticRoleCheck.Accept = function(self)
-  if not (AutomaticRoleCheck_Options.DisableOnce or (AutomaticRoleCheck.FirstCheck and AutomaticRoleCheck_Options.DisableOnceOnLogin)) then
-    self:Click()
+  if AutomaticRoleCheck_Options.Enabled and not (AutomaticRoleCheck_Options.DisableWhilstAFK and UnitIsAFK("player")) then
+    if not (AutomaticRoleCheck_Options.DisableOnce or (AutomaticRoleCheck.FirstCheck and AutomaticRoleCheck_Options.DisableOnceOnLogin)) then
+      self:Click()
+    end
+    if AutomaticRoleCheck.FirstCheck then
+      AutomaticRoleCheck.FirstCheck = false
+    end
+    if AutomaticRoleCheck_Options.DisableOnce then
+      AutomaticRoleCheck_Options.DisableOnce = false
+    end
   end
-  AutomaticRoleCheck.FirstCheck = false
-  AutomaticRoleCheck_Options.DisableOnce = false
 end
 
 LFGListApplicationDialog.SignUpButton:HookScript("OnShow", AutomaticRoleCheck.Accept)